requirements.txt
flask
piper-tts
pydub
cachetools
google-cloud-texttospeech
google-cloud-speech
transformers
torch